% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MediaWiki.R
\docType{data}
\name{bosox}
\alias{bosox}
\title{Wikipedia revision data for the Boston_Red_Sox page}
\format{a character vector of length 1}
\source{
was the result of a call to MediaWikiSpecialExport:
MediaWikiSpecialExport("Boston_Red_Sox", offset="2015-01-01T04:39:58Z", limit=3)
}
\usage{
bosox
}
\description{
An example XML output of Special:Export for the Boston_Red_Sox page.
}
\keyword{datasets}

