% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MediaWiki.R
\name{MediaWikiSpecialExport}
\alias{MediaWikiSpecialExport}
\title{Wrapper function for Wikipedia Special:Export}
\usage{
MediaWikiSpecialExport(pages, offset = 1, limit = 5)
}
\arguments{
\item{pages}{a character vector of pages to retrieve}

\item{offset}{the version with which to start}

\item{limit}{the number of versions to return}
}
\value{
the XML content returned
}
\description{
For repeated use of the Special:Export functionality in the WikiMedia API
specifically for Wikipedia, it made sense to have a convenience function.
}
\examples{
\dontrun{
bosox <- MediaWikiSpecialExport("Boston_Red_Sox", offset="2015-01-01T04:39:58Z", limit=3)
small <- MediaWikiSpecialExport("small", offset="2015-01-01T00:00:00Z", limit=6)
}
nchar(bosox)
nchar(small)
}

