% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MediaWiki.R
\name{MediaWikiSpecialExportParser}
\alias{MediaWikiSpecialExportParser}
\title{Data extraction class}
\usage{
MediaWikiSpecialExportParser()
}
\value{
a parser object
}
\description{
Given the standard format of the Special:Export API function, this object
can parse the results and produce a \code{data.frame}.
}
\details{
This class was adapted from my answers to an earlier assignment in the
course.
}
\examples{
\dontrun{
parser <- MediaWikiSpecialExportParser()
handler <- XML::xmlEventParse(file=small, branches=parser$saxHandler(), asText=TRUE)
revisions <- parser$details()
str(revisions)
revisions
}
}

