% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plotEvolution.R
\name{plot.evolution}
\alias{plot.evolution}
\title{Create an evolution visualization}
\usage{
\method{plot}{evolution}(x, ..., startAtTop = TRUE, interactive = TRUE,
  age = TRUE, colorPal = c("topo", "rainbow", "warm"), xtitle = NULL)
}
\arguments{
\item{x}{The object of type \code{evolution}.}

\item{...}{Additional parameters to ggvis.}

\item{startAtTop}{A logical scalar. Render starting at top of y axis?
This may be preferred for document visualization, so that visually it
grows \emph{from the top down}.}

\item{interactive}{A logical scalar. Render interactive features?
When TRUE, tooltips showing the class of content are shown on mouseover.
When FALSE a static image will be rendered.}

\item{age}{A logical scalar. When TRUE, colors are \emph{aged} over
time, by getting darker as they progess along the x axis.}

\item{colorPal}{A character scalar, either \code{topo}, \code{rainbow},
or \code{warm} which specifies the range of hues in the HSV color model.}

\item{xtitle}{A character scalar.  Title shown on the x axis.  If not
supplied, a title will be created using the evolution formula terms.}
}
\value{
Returns a \code{ggvis} visualization object.
}
\description{
It is the evolution dispatch method for the internal generic \code{plot}
}
\details{
\code{plot(evolutionObj, ...} or \code{plot.evolution(...)} produces a
visualization object of type \code{ggvis}.  It can be passed on to
other operations in the \code{ggvis} package, such as
\code{ggvis::view_dynamic(...)} and  \code{ggvis::add_legend(...)}.

Use code like \code{ggvis::view_dynamic(vis, port = 3838)} to visualize
interactivity in external local browser at \code{http://localhost:3838/}
}
\examples{
\dontrun{
evo <- evolution(texts ~ v(rev) + author, data=data, diff.fun=textDiff)
vis <- plot(evo)
view_dynamic(vis, port = 3838)
}
}

