% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MediaWiki.R
\docType{data}
\name{small}
\alias{small}
\title{Wikipedia revision data for the Small page}
\format{a character vector of length 1}
\source{
was the result of a call to MediaWikiSpecialExport:
MediaWikiSpecialExport("small", offset="2015-01-01T00:00:00Z", limit=6)
}
\usage{
small
}
\description{
An example XML output of Special:Export for the Small page.
}
\keyword{datasets}

