% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tempTest.R
\name{testPlot}
\alias{testPlot}
\title{Test R file for package build process}
\usage{
testPlot(col = c("red", "green", "blue"))
}
\arguments{
\item{col}{The color of points. Must be "red", "green" or "blue".  Default is "red".}
}
\value{
Output the plot
}
\description{
Test Plot
}
\examples{
\dontrun{testPlot("blue")}
}

